+++ lib/ex_unit/lib/ex_unit/runner.ex
@@ -1,29 +1,35 @@
 defmodule ExUnit.Runner do
   @moduledoc false
 
+  alias ExUnit.EventManager, as: EM
   @stop_timeout 30_000
 
-  defrecord Config, max_cases: 4, taken_cases: 0, async_cases: [], formatters: [],
-                    sync_cases: [], include: [], exclude: []
+  defrecord Config, max_cases: 4, taken_cases: 0, async_cases: [],
+                    sync_cases: [], include: [], exclude: [], manager: []
 
-  def run(async, sync, opts, load_us, return // :failures) do
-    opts   = normalize_opts(opts)
-    config = Config[max_cases: :erlang.system_info(:schedulers_online)]
+  def run(async, sync, opts, load_us) do
+    opts = normalize_opts(opts)
+
+    { :ok, pid } = EM.start_link
+    formatters = [ExUnit.RunnerStats|opts[:formatters]]
+    Enum.each formatters, &(:ok = EM.add_handler(pid, &1, opts))
+
+    config = Config[max_cases: :erlang.system_info(:schedulers_online), manager: pid]
     config = wrap_filters(config.update(opts))
-
-    :ok = ExUnit.Formatter.Manager.add_handler(ExUnit.Runner.FailureCounter, [])
 
     { run_us, _ } =
       :timer.tc fn ->
-        ExUnit.Formatter.Manager.suite_started(opts)
+        EM.suite_started(config.manager, opts)
         loop config.async_cases(async).sync_cases(sync)
       end
 
-    ExUnit.Formatter.Manager.suite_finished(run_us, load_us)
-    ExUnit.Formatter.Manager.call(ExUnit.Runner.FailureCounter, { :stop, return }, @stop_timeout)
+    EM.suite_finished(config.manager, run_us, load_us)
+    EM.call(config.manager, ExUnit.RunnerStats, :stop, @stop_timeout)
   end
 
   defp normalize_opts(opts) do
+    opts = Keyword.put_new(opts, :color, IO.ANSI.terminal?)
+
     if opts[:trace] do
       Keyword.put_new(opts, :max_cases, 1)
     else
@@ -87,7 +93,7 @@ defmodule ExUnit.Runner do
 
   defp run_case(config, pid, case_name) do
     ExUnit.TestCase[] = test_case = case_name.__ex_unit__(:case)
-    ExUnit.Formatter.Manager.case_started(test_case)
+    EM.case_started(config.manager, test_case)
 
     # Prepare tests, selecting which ones should
     # run and which ones were skipped.
@@ -104,7 +110,7 @@ defmodule ExUnit.Runner do
     # but we do send the notifications to formatter and other
     # entities involved.
     Enum.each failed, &run_test(config, &1, [])
-    ExUnit.Formatter.Manager.case_finished(test_case)
+    EM.case_finished(config.manager, test_case)
     send pid, { self, :case_finished, test_case }
   end
 
@@ -164,13 +170,13 @@ defmodule ExUnit.Runner do
   end
 
   defp run_test(config, test, context) do
-    ExUnit.Formatter.Manager.test_started(test)
+    EM.test_started(config.manager, test)
 
     if nil?(test.state) do
       test = spawn_test(config, test, Keyword.merge(test.tags, context))
     end
 
-    ExUnit.Formatter.Manager.test_finished(test)
+    EM.test_finished(config.manager, test)
   end
 
   defp spawn_test(_config, test, context) do
@@ -253,51 +259,5 @@ defmodule ExUnit.Runner do
   # All other cases
   defp prune_stacktrace([h|t]), do: [h|prune_stacktrace(t)]
   defp prune_stacktrace([]), do: []
-
-  # Small helper to count all the failures
-  defmodule FailureCounter do
-    @moduledoc false
-
-    use GenEvent.Behaviour
-
-    def init(_opts) do
-      { :ok, { 0, 0 } } # { total, failures }
-    end
-
-    def handle_call({ :stop, :failures }, { _, fails }) do
-      { :remove_handler, fails }
-    end
-    def handle_call({ :stop, :total }, { total, fails }) do
-      { :remove_handler, total }
-    end
-    def handle_call({ :stop, tup }, { total, fails }) when is_tuple(tup) do
-      rv = tuple_to_list(tup)
-           |> Enum.map(fn
-               :total -> total
-               :failures -> fails
-              end)
-           |> list_to_tuple()
-      { :remove_handler, rv }
-    end
-
-    def handle_event({ :case_finished,  ExUnit.TestCase[state: { :failed, _ }] }, { total, fails }) do
-      { :ok, { total, fails + 1 } }
-    end
-
-    def handle_event({ :test_finished,  ExUnit.Test[state: { :failed, _ }] }, { total, fails }) do
-      { :ok, { total + 1, fails + 1 } }
-    end
-
-    def handle_event({ :test_finished, ExUnit.Test[state: { :skip, _ }] }, { total, fails }) do
-      { :ok, { total, fails } }
-    end
-
-    def handle_event({ :test_finished, _ }, { total, fails }) do
-      { :ok, { total + 1, fails } }
-    end
-    def handle_event(e, { total, fails }) do
-      { :ok, { total, fails } }
-    end
-  end
 end